services/intelligence-worker/pyproject.toml
[project]
name = "benevolent-director-intelligence-worker"
version = "0.0.0"
description = "v2 intelligence worker scaffold"
requires-python = ">=3.12"
dependencies = [
  "beautifulsoup4>=4.12",
  "google-cloud-storage>=2.18",
  "httpx>=0.27",
  "pdfplumber>=0.11",
  "psycopg[binary]>=3.2",
  "pydantic>=2.10",
  "structlog>=24.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.4.0",
  "mypy>=1.10",
]

[project.scripts]
intelligence-worker = "intelligence_worker.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/intelligence_worker"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
testpaths = ["tests"]
